options (de)activation in options view
