+++ lib/widgets/dialogs/delete_confirmation_dialog.dart
@@ -1,0 +1,39 @@
+import 'package:flutter/material.dart';
+
+class DeleteConfirmationDialog extends StatelessWidget {
+  final String title;
+  final String message;
+  final VoidCallback onConfirmed;
+
+  const DeleteConfirmationDialog({
+    super.key,
+    required this.title,
+    required this.message,
+    required this.onConfirmed,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: Text(title),
+      content: Text(message),
+      actions: [
+        TextButton(
+          onPressed: () => Navigator.of(context).pop(),
+          child: Text('Cancel'),
+        ),
+        ElevatedButton(
+          onPressed: () {
+            onConfirmed();
+            Navigator.of(context).pop();
+          },
+          style: ElevatedButton.styleFrom(
+            backgroundColor: Theme.of(context).colorScheme.error,
+            foregroundColor: Theme.of(context).colorScheme.onError,
+          ),
+          child: Text('Delete'),
+        ),
+      ],
+    );
+  }
+}
